coach_service/requirements.txt
flask==2.2.3
Flask-Cors==3.0.10
requests==2.26.0
openai==1.73.0
python-dotenv==0.19.0
redis==4.0.2
prometheus-client==0.17.1